refactor middlewares in their own folder, move buildFiles to libs/buildFiles
