+++ website/src/middlewares/cors.js
@@ -1,0 +1,7 @@
+module.exports = function(req, res, next) {
+  res.header("Access-Control-Allow-Origin", req.headers.origin || "*");
+  res.header("Access-Control-Allow-Methods", "OPTIONS,GET,POST,PUT,HEAD,DELETE");
+  res.header("Access-Control-Allow-Headers", "Content-Type,Accept,Content-Encoding,X-Requested-With,x-api-user,x-api-key");
+  if (req.method === 'OPTIONS') return res.send(200);
+  return next();
+};
