+++ website/src/routes/dataexport.js
@@ -4,13 +4,13 @@ var dataexport = require('../controllers/dataexport');
 var auth = require('../controllers/auth');
 var nconf = require('nconf');
 var i18n = require('../i18n');
-var middleware = require('../middleware.js');
+var locals = require('../middlewares/locals');
 
 /* Data export */
 router.get('/history.csv',auth.authWithSession,i18n.getUserLanguage,dataexport.history); //[todo] encode data output options in the data controller and use these to build routes
 router.get('/userdata.xml',auth.authWithSession,i18n.getUserLanguage,dataexport.leanuser,dataexport.userdata.xml);
 router.get('/userdata.json',auth.authWithSession,i18n.getUserLanguage,dataexport.leanuser,dataexport.userdata.json);
-router.get('/avatar-:uuid.html', i18n.getUserLanguage, middleware.locals, dataexport.avatarPage);
-router.get('/avatar-:uuid.png', i18n.getUserLanguage, middleware.locals, dataexport.avatarImage);
+router.get('/avatar-:uuid.html', i18n.getUserLanguage, locals, dataexport.avatarPage);
+router.get('/avatar-:uuid.png', i18n.getUserLanguage, locals, dataexport.avatarImage);
 
 module.exports = router;
